+++ hyperspy/tests/component/test_component_print_current.py
@@ -18,9 +18,11 @@
 import pytest
 
 from hyperspy.datasets.example_signals import EDS_SEM_Spectrum
-from hyperspy.misc.model_tools import (_format_string,
-                                       current_component_values,
-                                       current_model_values)
+from hyperspy.misc.model_tools import (
+    _format_string,
+    CurrentComponentValues,
+    CurrentModelValues,
+    )
 
 
 class TestSetParameters:
@@ -40,8 +42,8 @@ class TestSetParameters:
     @pytest.mark.parametrize("only_free, only_active", [(True, False), (True, False)])
     def test_component_current_component_values(self, only_free, only_active):
         "Many decimals aren't printed, few decimals are"
-        string_representation = str(current_component_values(self.component, only_free, only_active).__repr__())
-        html_representation = str(current_component_values(self.component, only_free, only_active)._repr_html_())
+        string_representation = str(CurrentComponentValues(self.component, only_free, only_active).__repr__())
+        html_representation = str(CurrentComponentValues(self.component, only_free, only_active)._repr_html_())
         assert "1.234" in string_representation
         assert "1.23456789012" not in string_representation
         assert "1.234" in html_representation
@@ -49,16 +51,16 @@ class TestSetParameters:
 
     def test_component_current_component_values_only_free(self, only_free=True, only_active=False):
         "Parameters with free=False values should not be present in repr"
-        string_representation = str(current_component_values(self.component_not_free, only_free, only_active).__repr__())
-        html_representation = str(current_component_values(self.component_not_free, only_free, only_active)._repr_html_())
+        string_representation = str(CurrentComponentValues(self.component_not_free, only_free, only_active).__repr__())
+        html_representation = str(CurrentComponentValues(self.component_not_free, only_free, only_active)._repr_html_())
         assert "9.87" not in string_representation
         assert "9.87" not in html_representation
 
     @pytest.mark.parametrize("only_free, only_active", [(True, False), (True, False)])
     def test_component_current_model_values(self, only_free, only_active):
         "Many decimals aren't printed, few decimals are"
-        string_representation = str(current_model_values(self.model, only_free, only_active).__repr__())
-        html_representation = str(current_model_values(self.model, only_free, only_active)._repr_html_())
+        string_representation = str(CurrentModelValues(self.model, only_free, only_active).__repr__())
+        html_representation = str(CurrentModelValues(self.model, only_free, only_active)._repr_html_())
         assert "1.234" in string_representation
         assert "1.23456789012" not in string_representation
         assert "1.234" in html_representation
@@ -76,8 +78,8 @@ class TestSetParameters:
     @pytest.mark.parametrize("only_free, only_active", [(True, False), (True, False)])
     def test_component_current_model_values_comp_list(self, only_free, only_active):
         comp_list = [self.component, self.component_not_free, self.component_inactive]
-        string_representation = str(current_model_values(self.model, only_free, only_active, comp_list).__repr__())
-        html_representation = str(current_model_values(self.model, only_free, only_active, comp_list)._repr_html_())
+        string_representation = str(CurrentModelValues(self.model, only_free, only_active, comp_list).__repr__())
+        html_representation = str(CurrentModelValues(self.model, only_free, only_active, comp_list)._repr_html_())
         assert "1.234" in string_representation
         assert "1.23456789012" not in string_representation
         assert "1.234" in html_representation
@@ -90,28 +92,25 @@ class TestSetParameters:
             assert "5.67" not in string_representation
             assert "5.67" not in html_representation
 
-    @pytest.mark.parametrize("fancy", (True, False))
-    def test_model_current_model_values(self, fancy):
-        self.model.print_current_values(fancy=fancy)
+    def test_model_current_model_values(self):
+        self.model.print_current_values()
 
-    @pytest.mark.parametrize("fancy", (True, False))
-    def test_component_print_current_values(self, fancy):
-        self.model[0].print_current_values(fancy=fancy)
+    def test_component_print_current_values(self):
+        self.model[0].print_current_values()
 
-    @pytest.mark.parametrize("fancy", (True, False))
-    def test_model_print_current_values(self, fancy):
-        self.model.print_current_values(fancy=fancy)
+    def test_model_print_current_values(self):
+        self.model.print_current_values()
 
-    def test_zero_in_fancy_print(self):
+    def test_zero_in_html_print(self):
         "Ensure parameters with value=0 are printed too"
-        assert "<td>a1</td><td>True</td><td>     0</td>" in current_component_values(self.model[0])._repr_html_()
+        assert "<td>a1</td><td>True</td><td>     0</td>" in CurrentComponentValues(self.model[0])._repr_html_()
 
     def test_zero_in_normal_print(self):
         "Ensure parameters with value=0 are printed too"
-        assert "            a0 |    True |          0 |" in str(current_component_values(self.model[0]).__repr__)
+        assert "            a0 |    True |          0 |" in str(CurrentComponentValues(self.model[0]).__repr__)
 
     def test_twinned_in_print(self):
-        assert "             A | Twinned |" in str(current_component_values(self.model[2]).__repr__()).split('\n')[4]
+        assert "             A | Twinned |" in str(CurrentComponentValues(self.model[2]).__repr__()).split('\n')[4]
 
     def test_related_tools(self):
         assert _format_string(None) == ""